CLOUDERA-BUILD Update upgrade script for CDH use
The original script would upgrade Kudu parcels, wait for them to
install, and finally restart the Kudu service. This patch tweaks the
behavior so the parcel name can be user specified, and the service name
can specify any service, rather than just Kudu services. If the service
is not specified, all stale services are now restarted.

Also included are a 'port' flag, and a 'dry-run' flag, which prints
messages useful for testing without actually upgrading the cluster.