Add additonal form and tests for GetConsensusRole()
There are cases where calling GetConsensusRole() without creating a
ConsensusStatePB object would be more efficient. This patch adds a
version of GetConsensusRole() with the following signature:

  RaftPeerPB::Role GetConsensusRole(const std::string& peer_uuid,
                                    const std::string& leader_uuid,
                                    const RaftConfigPB& config);

This patch also fixes a bug where the role of a non-voter that was
thought to be the leader would be returned as NON_PARTICIPANT instead of
LEARNER. However, that is an illegal state so it should not actually
affect production code.

Also added a unit test for both forms of GetConsensusRole().

Change-Id: Ic28f15c40ce99f89481cb6869d4843cbc024ed48
Reviewed-on: http://gerrit.cloudera.org:8080/9260
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit c8a1c0f977c122e4da0e12deec9e23f7ce6a9e74)