[consensus] add queue/cmeta consistency DCHECK
Added an assertion on the consistency of the consensus metadata and
the consensus queue (effective for DEBUG builds only).

This is in the context of tracking and fixing the bug we hit earlier,
and we guess it happened because of the issue described in KUDU-2274.

Change-Id: Id9c7a16add29c8ab0228085c2009ee17a435d3e8
Reviewed-on: http://gerrit.cloudera.org:8080/9272
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
(cherry picked from commit 194fd8b169f29aafbd78a47709ac51d2e8354a1a)