python: fix build on Python 2.6
We appear to have been bitten by a bug [1] in setuptools_scm. Here's the
recommended workaround.

For the life of me, I couldn't isolate this to any particular package
version change. I tried "last known good" versions of pip, setuptools,
setuptools_scm, pytest, and pytest-runner. In every case the exception
persisted when building on CentOS 6.6 with Python 2.6.6, but not on Ubuntu
16.04 with Python 2.7.12. My best guess is that it's due to the Python
version, but I don't know that for sure.

1. https://github.com/pypa/setuptools_scm/issues/209

Change-Id: I0cf7f64363b308e06ac9bcdebcfc3832ff7b65b9
Reviewed-on: http://gerrit.cloudera.org:8080/9468
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>