KUDU-2342. consensus: use tighter bound for non-voter promotion
This patch replaces the NON_VOTER promotion heuristic based on the
--consensus_promotion_max_wal_entries_behind gflag with a heuristic that
roughly matches the promotion criteria outlined in Diego Ongaro's
dissertation.

The approach is that we track the operation count of the last batch
successfully sent from leader to peer. If the peer is within that number
of ops of the commit index, the peer is eligible for promotion.

The aforementioned gflag has been removed.

This patch passes existing tests and did well on manual testing. I will
follow up on this patch with an automated test.

Change-Id: Iff517f01d6dc25eb15d01593dd57b7dc0dd25956
Reviewed-on: http://gerrit.cloudera.org:8080/9627
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit f2479e21d5a3002ebf5b1012fde83a6cffc2db82)

Conflicts:
  src/kudu/consensus/consensus_queue.cc