KUDU-2379. java: exportAuthenticationCredentials must connect to the cluster
In the case that exportAuthenticationCredentials is called before the
client has connected to the cluster, it needs to trigger a connection in
order to obtain credentials. KUDU-2259 broke this when it changed
exportAuthenticationCredentials() to return a non-null credential in
this case.

The fix just tracks whether the client has successfully connected to a
cluster, and if it has not, makes sure that it has done so before
generating credentials.

Tested manually using spark2-submit against a secure cluster. Prior to
this fix, it did not succeed. A new unit test is included as well.

Change-Id: Icd1b69abadb4b9c97f6c8f7ed38897c33ca7ff72
Reviewed-on: http://gerrit.cloudera.org:8080/9814
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>
(cherry picked from commit aa94d16fda6cd6f3f413decf46bf9353680480c7)