KUDU-2427: make ksck-test replica order deterministic
On Ubuntu 18.04, the iteration order of std::unordered_map isn't the same as
on other distros. Behavioral changes like this are to be expected as the
iteration order is implementation-specific. However, we had some code in
ksck-test that expected a specific iteration order. Let's fix this by
switching from an std::unordered_map to an std::map.

==C5_FEATURE_KUDU_TABLET== @ CDH-71222

Change-Id: I5da71f3582257febaccae7f142e5eb3a22548c19
Reviewed-on: http://gerrit.cloudera.org:8080/10433
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
(cherry picked from commit 5aca86284967770048da6440b54ce0daea2d18eb)