RELENG-3431: pre-commit hook for C6 commit flow
This patch adds a script that can be called by C6 pre-commit jobs. The
script also pulls in the requirements to run Kudu with the dist test
infrastructure.

It's worth noting that these builds are a bit flaky, but it's still
worth committing so we have a base with which to merge patches with
confidence.

Change-Id: I226d1cc6f15f5df3d2f7770c9f0083fc201bebef