[rebalance-test] fix running the test on Ubuntu 18.04
Updated custom equality operators for TableBalanceInfo and
ClusterBalanceInfo.  With this update, the operator compares just the
contents of corresponding sub-fields of the ServersByCountMap type
(which is map<int32_t, string>), disregarding the order of elements for
the same key.  That's the exact behavior desired: the ordering of
elements in the ServersByCountMap is not important in all scenarios
of the rebalance-test.

The motivation for this change is the fact that the test fails on
Ubuntu 18.04 without this patch.  Implicitly, the reference results
relied on the order of elements in hash dictionary container used in
the implementation of the Rebalancer::KsckResultsToClusterBalanceInfo()
method.  It seems the implementation of std::unordered_map in libstdc++
shipped with Ubuntu 18.04 has changed compared with older Linux distro
versions.

==C5_FEATURE_KUDU_TABLET== @ CDH-71222

Change-Id: Ie4dc5d22bab86f6ab61f18f05e45a063bbf38eba
Reviewed-on: http://gerrit.cloudera.org:8080/10812
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
(cherry picked from commit 9c89c2a9aad6956c39f04454467789fa6d8288ef)