spark: Expose socketReadTimeoutMs to Spark connector
This patch exposes socketReadTimeoutMs in the KuduContext and the
DefaultSource.

This patch also performs a bit of cleanup by renaming
the KuduConnection object to KuduClientCache, which seems like a more
appropriate name.

Because socketReadTimeout is a KuduClient configuration parameter
related to connection handling, socketReadTimeout was incorporated into
the client cache key.

Manually tested in spark-shell using spark-on-yarn.

Added a basic test to ensure that the parameter is properly parsed by
the DefaultSource and configured in the KuduRelation instance.

==C5_APPROVED_BUGFIX== @ CDH-70764

Change-Id: I0ab0ff0b242790caffb7e2848958148ffe547c4d
Reviewed-on: http://gerrit.cloudera.org:8080/10839
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>