[tools] ksck improvements [1/n]: Eliminate KsckMaster class
ksck has a KsckCluster and a KsckMaster class. The KsckCluster class's
methods more-or-less map one-to-one with KsckMaster class methods.
The KsckMaster class represents the collective masters of a cluster,
not an individual master.

This patch combines the two classes into a KsckCluster class. This
class is the base class for a MockKsckCluster class and a
RemoteKsckCluster class, for tests and "real life", respectively.

This is just refactoring. There are no functional changes in this patch.

==C5_FEATURE_KUDU_TABLET== @ CDH-71222

Change-Id: I21f9e244b6ba10e11327c744abc1ac72a9b2ab1c
Reviewed-on: http://gerrit.cloudera.org:8080/9881
Reviewed-by: Attila Bukor <abukor@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>
(cherry picked from commit 7e35aee9626bf3370d9b1d89fd1f3440b47ff7a8)