[tools] enable RF=1 replica movement
Regardless of the fact that Kudu upstream version for CDH5.16.x is
reported as 1.7.0, Kudu in CDH5.16+ contains fixes for KUDU-2443 that
allows for moving of single replica tablets.  To reflect that, this
patch updates VersionSupportsRF1Movement() utility function to
parse CDH version out of the string version and compare it with
5.16.0 as the minimum version that supports RF=1 replica movement.

This patch also fixes CDH-72918.

==C5_FEATURE_KUDU_TABLET== @ CDH-71222