[raft_consensus-itest] fix flake in Test_KUDU_1735
Fixed flake in RaftConsensusParamReplicationModesITest.Test_KUDU_1735
scenario of the raft_consensus-itest suite, for both the 3-4-3 and 3-2-3
replica management schemes.

The flakiness was due to a few reasons, where the most prominent one was
the race in 3-4-3 mode between shutting down replica's consensus
upon DeleteTablet(TABLET_DATA_TOMBSTONED) and re-discovering the peer
by the leader once the catalog manager added the replica back as a
non-voter.  Another minor flake was due to not waiting for the manually
elected leader to commit an operation on its own term.  And the one
was due to waiting for RECEIVED_OPID instead of COMMITTED_OPID Raft
operation index before setting the --fault_crash_before_append_commit
flag to 1.

I ran the RaftConsensusParamReplicationModesITest.Test_KUDU_1735
scenario using dist_test with --stress_cpu_threads=16 before and after
the fix with the following results:

  before the fix (950 out of 1024 failed):
    http://dist-test.cloudera.org/job?job_id=aserbin.1542868617.80411

  after the fix  (  0 out of 1024 failed):
    http://dist-test.cloudera.org/job?job_id=aserbin.1542871109.113847

Change-Id: If44ad0e8363a3aead409484cff68843f1e5d6b6d
Reviewed-on: http://gerrit.cloudera.org:8080/11981
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>