OPSAPS-48407 Remove full/failed data dirs health tests for now
For reasons that are unclear at this time, CM considers the health of
one or more Kudu role instances to be bad because there is not enough
data to test for the full and failed data dirs tests, despite the fact
that all indications are that CM is gathering the metrics data from Kudu
just fine.

Until there's deeper understanding and a fix for the problem on the CM
side, let's get rid of these health tests.

Change-Id: I68c4bdbc9f452e4d6de18eaa75ae8d4add7fd803
(cherry picked from commit 7e735035511d2176189e5fee6879d969dc8727d4)